+++ inc/lilyParser/lilyTerms.ts
@@ -1,7 +1,7 @@
 
 import _ from "lodash";
 
-import {WHOLE_DURATION_MAGNITUDE, FractionNumber, lcmMulti} from "./utils";
+import {WHOLE_DURATION_MAGNITUDE, FractionNumber, lcmMulti, gcd} from "./utils";
 import * as idioms from "./idioms";
 import {LILYPOND_VERSION} from "../constants";
 import * as measureLayout from "../lilyNotation/measureLayout";
@@ -1325,12 +1325,49 @@ export class MusicBlock extends BaseTerm {
 
 
 	// with side effects
-	redivide ({recursive = true} = {}) {
+	redivide ({recursive = true, measureHeads = null}: {recursive?: boolean, measureHeads?: number[]} = {}) {
 		if (recursive) {
 			this.forEachTerm(MusicBlock, block => {
 				if (!block._parent || block._parent.cmd !== "alternative")
-					block.redivide();
+					block.redivide({recursive, measureHeads});
 			});
+		}
+
+		// split rests
+		if (measureHeads) {
+			this.body = [].concat(...this.body.map(term => {
+				if (!(term instanceof Rest) || term.name !== "s" || !Number.isInteger(term._measure))
+					return [term];
+
+				const nextHead = measureHeads[term._measure];
+				const endTick = term._tick + term.durationMagnitude;
+				if (nextHead > 0 && endTick > nextHead) {
+					const post_events = term.post_events;
+
+					let startTick = term._tick;
+					const rests = [];
+					let nextMeasure;
+					for (nextMeasure = term._measure; nextMeasure < measureHeads.length && endTick > measureHeads[nextMeasure]; ++nextMeasure) {
+						const rest = new Rest({name: "s", duration: Duration.fromMagnitude(measureHeads[nextMeasure] - startTick), post_events: []});
+						rest._measure = nextMeasure;
+						rest._lastMeasure = nextMeasure;
+						rests.push(rest);
+
+						startTick = measureHeads[nextMeasure];
+					}
+
+					const rest = new Rest({name: "s", duration: Duration.fromMagnitude(endTick - startTick), post_events: [...post_events]});
+					rest._measure = nextMeasure;
+					rest._lastMeasure = nextMeasure;
+					rests.push(rest);
+
+					console.assert(rests.reduce((sum, rest) => sum + rest.durationMagnitude, 0) === term.durationMagnitude, "duration splitting error:", rests, term);
+
+					return rests;
+				}
+
+				return [term];
+			}));
 		}
 
 		const isPostTerm = term => !term
@@ -2104,6 +2141,34 @@ export class Duration extends BaseTerm {
 			Duration._default = new Duration({number: 4, dots: 0});
 
 		return Duration._default;
+	}
+
+
+	static fromMagnitude (magnitude: number): Duration {
+		if (!Number.isInteger(magnitude)) {
+			console.warn("magnitude must be integer:", magnitude);
+			return null;
+		}
+
+		const di = gcd(magnitude, WHOLE_DURATION_MAGNITUDE);
+		const denominator = WHOLE_DURATION_MAGNITUDE / di;
+		const numerator = magnitude / di;
+		if (!Number.isInteger(Math.log2(denominator)))
+			return new Duration({number: 1, dots: 0, multipliers: [`${numerator}/${denominator}`]});
+
+		switch (numerator) {
+		case 1:
+			return new Duration({number: denominator, dots: 0});
+
+		case 3:
+			return new Duration({number: denominator / 2, dots: 1});
+
+		case 7:
+			return new Duration({number: denominator / 4, dots: 2});
+
+		default:
+			return new Duration({number: denominator, dots: 0, multipliers: [numerator.toString()]});
+		}
 	}